Mess with the initial connection more

+++ src/Universalis.Entities/Uploaders/UploadApplication.cs
@@ -2,7 +2,7 @@
 {
     public class UploadApplication
     {
-        public byte Id { get; set; }
+        public uint Id { get; set; }
 
         public string Name { get; set; }
     }
